+++ CHANGELOG.md
@@ -11,6 +11,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Deprecated
 #### Removed
 
+## [0.3.0] - 2025-01-15
+
+### Added
+
+* feat: Add unit tests
+
+### Changed
+
+* feat: Do not repeat rule name on each rule result when they are grouped by rule name
+* feat: Ignore license headers rule when it is an empty array
+
+### Fixed
+
+* chore: Remove unused script from package.json
+
 ## [0.2.0] - 2025-01
 
 ### Added